# openapi.json
{
  "components": {
    "schemas": {
...
      "EntitiesResponse": {
        "additionalProperties": true,
        "properties": {
          "entities": {
            "items": {
              "$ref": "#/components/schemas/EntitySummary"
            },
            "title": "Entities",
            "type": "array"
          },
          "matches": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Matches",
            "type": "array"
          }
        },
        "title": "EntitiesResponse",
        "type": "object"
      },
...
      "EntityFilterOptions": {
        "additionalProperties": true,
        "properties": {
          "entity_types": {
            "items": {
              "type": "string"
            },
            "title": "Entity Types",
            "type": "array"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "title": "Sources",
            "type": "array"
          },
          "taxonomy_ids": {
            "items": {
              "type": "string"
            },
            "title": "Taxonomy Ids",
            "type": "array"
          }
        },
        "title": "EntityFilterOptions",
        "type": "object"
      },
...
      "EntityPksRequest": {
        "additionalProperties": true,
        "properties": {
          "entityPks": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Entitypks",
            "type": "array"
          },
          "pks": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Pks",
            "type": "array"
          }
        },
        "title": "EntityPksRequest",
        "type": "object"
      },
      "EntityPublicIdsRequest": {
        "additionalProperties": true,
        "properties": {
          "publicIds": {
            "items": {
              "type": "string"
            },
            "title": "Publicids",
            "type": "array"
          },
          "public_ids": {
            "items": {
              "type": "string"
            },
            "title": "Public Ids",
            "type": "array"
          }
        },
        "title": "EntityPublicIdsRequest",
        "type": "object"
      },
      "EntityResolveRequest": {
        "additionalProperties": true,
        "properties": {
          "identifiers": {
            "items": {
              "type": "string"
            },
            "title": "Identifiers",
            "type": "array"
          }
        },
        "title": "EntityResolveRequest",
        "type": "object"
      },
...
      "EntitySearchResponse": {
        "additionalProperties": true,
        "properties": {
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count"
          },
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Ms"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/EntitySummary"
            },
            "title": "Entities",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntitySearchCursor"
              },
              {
                "type": "null"
              }
            ]
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/EntitySearchRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "title": "EntitySearchResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": true,
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "GroupRelationshipsRequest": {
        "additionalProperties": true,
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/SearchFilters"
          },
          "group_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Key"
          },
          "limit": {
            "default": 50,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "member_keys": {
            "items": {
              "type": "string"
            },
            "title": "Member Keys",
            "type": "array"
          },
          "offset": {
            "default": 0,
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "query": {
            "default": "",
            "title": "Query",
            "type": "string"
          },
          "resources": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resources"
          }
        },
        "title": "GroupRelationshipsRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "REST API for querying OmniPath knowledge graphs directly from nested Parquet files via DuckDB. The Svelte explorer uses `/api/*` (and `/app-api/*` as an application alias). Interactive docs: `/api/docs`.",
    "title": "OmniPath Parquet API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/entities/by-pks": {
      "post": {
        "description": "Hydrate entities by primary key.",
        "operationId": "entities_by_pks_entities_by_pks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityPksRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitiesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Entities By Pks",
        "tags": [
          "entities"
        ]
      }
    },
    "/entities/by-public-ids": {
      "post": {
        "description": "Hydrate entities by public identifier.",
        "operationId": "entities_by_public_ids_entities_by_public_ids_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityPublicIdsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitiesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Entities By Public Ids",
        "tags": [
          "entities"
        ]
      }
    },
    "/entities/examples": {
      "get": {
        "description": "Curated landing suggestions from the selected release; not a search page.",
        "operationId": "entity_examples_entities_examples_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitySearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Entity Examples",
        "tags": [
          "entities"
        ]
      }
    },
    "/entities/filter-options": {
      "get": {
        "description": "Distinct entity types, sources, and taxa for explorer filters.",
        "operationId": "entity_filter_options_entities_filter_options_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityFilterOptions"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Entity Filter Options",
        "tags": [
          "entities"
        ]
      }
    },
    "/entities/group-relationships": {
      "post": {
        "operationId": "group_relationships_entities_group_relationships_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupRelationshipsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Group Relationships",
        "tags": [
          "entities"
        ]
      }
    },
    "/entities/groups": {
      "post": {
        "description": "Group matching entities using an explicit grouping strategy.\n\nCounts and member pagination are scoped to the current query and filters.\nChemical connectivity uses the first 14 InChIKey characters. Other entities\nand compounds with no unambiguous key remain separate singleton results.",
        "operationId": "connectivity_groups_entities_groups_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityGroupsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Connectivity Groups",
        "tags": [
          "entities"
        ]
      }
    },
    "/entities/resolve": {
      "post": {
        "description": "Resolve raw identifiers to canonical entities.",
        "operationId": "resolve_entities_entities_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitiesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Resolve Entities",
        "tags": [
          "entities"
        ]
      }
    },
    "/entities/scoped-facets": {
      "post": {
        "description": "Facet counts for the current entity selection.",
        "operationId": "scoped_entity_facets_entities_scoped_facets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScopedEntityFacetsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FacetCount"
                  },
                  "title": "Response Scoped Entity Facets Entities Scoped Facets Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Scoped Entity Facets",
        "tags": [
          "entities"
        ]
      }
    },
    "/entities/search": {
      "get": {
        "description": "Autocomplete-style entity search used by GET callers.",
        "operationId": "search_entities_get_entities_search_get",
        "parameters": [
          {
            "description": "Identifier or label substring.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "description": "Identifier or label substring.",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "JSON object `{relationCount, entityPk}`.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JSON object `{relationCount, entityPk}`.",
              "title": "Cursor"
            }
          },
          {
            "description": "JSON object of search filters.",
            "in": "query",
            "name": "filters",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JSON object of search filters.",
              "title": "Filters"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitySearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Search Entities Get",
        "tags": [
          "entities"
        ]
      },
      "post": {
        "description": "Entity search with JSON filters, cursor pagination, and optional resource scope.",
        "operationId": "search_entities_post_entities_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntitySearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitySearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Search Entities Post",
        "tags": [
          "entities"
        ]
      }
    },
    "/entities/{entity_id}": {
      "get": {
        "description": "Fetch one entity by its exact entity key. No identifier or text fallback.",
        "operationId": "get_entity_entities__entity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "includeRelationships",
            "required": false,
            "schema": {
              "default": true,
              "title": "Includerelationships",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "detail_limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Detail Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "detail_offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Detail Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetailsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Entity",
        "tags": [
          "entities"
        ]
      }
    },
    "/entities/{entity_id}/relationships": {
      "get": {
        "operationId": "entity_relationships_entities__entity_id__relationships_get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Entity Relationships",
        "tags": [
          "entities"
        ]
      }
    },
...
    "/resources": {
      "get": {
        "description": "List indexed resources. `shape=svelte` returns the explorer catalog.",
        "operationId": "list_resources_resources_get",
        "parameters": [
          {
            "description": "Pass `svelte` for the explorer catalog shape.",
            "in": "query",
            "name": "shape",
            "required": false,
            "schema": {
              "default": "web",
              "description": "Pass `svelte` for the explorer catalog shape.",
              "title": "Shape",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourcesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Resources",
        "tags": [
          "resources"
        ]
      }
    },
...
  },
...
}